+++ cell_service.go
@@ -11,7 +11,7 @@ import (
 )
 
 const (
-	CurrentVersion = "1.2.0"
+	CurrentVersion = "1.3.0"
 )
 
 type CellService struct {